pyproject.toml
[tool.poetry]
name = "RandomPyDefis"
version = "0.1.0"
description = "Get a random challenge from pydefis"
authors = ["Evann DREUMONT <53308142+LeGmask@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"
"discord.py" = "^1.6.0"
python-dotenv = "^0.15.0"
humanize = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
